Correct flags when using gokrb5
Use the same consistent GSSAPI flags as when using the Apcera
implementation. Setting `gssapi.ContextFlagReplay` requires us to set
the sequence number correctly in the MIC tokens based on what the
authenticator is seeded with. Use the correct constant name for the
APOptions parameter; confusingly `gssapi.ContextFlagMutual` and
`flags.APOptionMutualRequired` have the same value so it was sort of
working even though the code didn't read intuitively.

Fixes #23